Imageboard.Actions.Auth

Documentation

modPage :: ActionM ()

If user is logged in, render his account page. Otherwise render login form.

tryLogin :: ActionM ()

Try to log in a user with supplied data from login form.

logout :: ActionM ()

Log out a user.

allowLoggedIn :: ActionM () -> ActionM ()

Allow any logged in user to perform an action.

allowAdmin :: ActionM () -> ActionM ()

Allow admin user to perform an action.

changePass :: ActionM ()

Try to change password of a user.