Imageboard.Database

Documentation

setupDb :: IO ()

Create sqlite3 database file using schema file.

Post and thread insertion

insertPost :: Board -> Int -> PostStub -> IO ()

Insert post with no file attached into database.

insertFile

Arguments

:: File

-> IO Int
File Id.

Insert file information into database.

insertPostWithFile

Arguments

:: Board

-> Int

-> PostStub

-> Int
fileId

-> IO ()
Post Id.

Insert post with file attached, using previously obtained file Id.

insertThread :: Board -> PostStub -> IO ()

Insert new thread with no file attached.

insertThreadWithFile :: Board -> PostStub -> Int -> IO ()

Insert post with file attached, using previously obtained file Id.

Post and thread queries

getPosts :: Int -> IO [Post]

getPosts n returns n most recent posts, from all boards.

getThread

Arguments

:: Board

-> Int
Post number.

-> IO (Maybe Thread)

Get thread by board name and number.

checkThread :: Board -> Int -> IO Bool

Check if thread with given number exists on given board.

getThreads :: Board -> IO [ThreadHead]

Get all threads from given board.

getThreadInfo :: Board -> Int -> IO (Maybe ThreadInfo)

Get thread information by board name and number.

getFileId

Arguments

:: Text
Filename

-> IO (Maybe Int)
File Id, if given filename exists in database.

Obtain ID of previously inserted file.

Board queries and management

getConstraints :: Board -> IO (Maybe BoardConstraints)

Get constraints of given board, if it exists.

getBoardNames :: IO [Board]

Get names of all existing boards.

getBoardInfo :: Board -> IO (Maybe BoardInfo)

Get more detailed information for given board.

getBoardInfos :: IO [BoardInfo]

Get more detailed information for all existing boards.

updateBoard

Arguments

:: Board
Old board name.

-> BoardInfo

-> BoardConstraints

-> IO ()

Change board configuration.

insertBoard :: BoardInfo -> BoardConstraints -> IO ()

Insert new board into a database.

Account management

getPasswordHash :: Username -> IO (Maybe (PasswordHash Bcrypt))

Get hashed password if given user exists.

insertSessionToken :: Username -> SessionKey -> IO ()

Insert new sesion token for given user. Will delete old session tokens for the user.

removeSessionToken :: SessionKey -> IO ()

Remove given session token from database.

getAccounts :: IO [AccountInfo]

Get list of all accounts.

getAccountByToken :: SessionKey -> IO (Maybe AccountInfo)

Get account associated with given session token, if it exists.

checkAccount :: Username -> IO Bool

Check if account with given username exists.

checkSession :: SessionKey -> IO Bool

Check if session token is still valid.

insertAccount :: Username -> PasswordHash Bcrypt -> Role -> IO ()

Insert new account into database.

changePassword :: Username -> PasswordHash Bcrypt -> IO ()

Change password of given user.

Moderation

removeFile :: Text -> IO ()

Remove file with given filename from database.

removePost :: Board -> Int -> IO ()

Remove posts with given number from given board.

unlinkFile :: Board -> Int -> IO ()

Unlink file from post with given number on given board, a.k.a. turn it into post with no file attached, without removing file from database.

Toggle one of ThreadFlags

toggleSticky :: Board -> Int -> IO ()

toggleLock :: Board -> Int -> IO ()

toggleAutosage :: Board -> Int -> IO ()

toggleCycle :: Board -> Int -> IO ()

Captcha

insertCaptcha :: String -> IO ()

Store new captcha in database. Delete expired captchas.

checkCaptcha :: Text -> IO Bool

Check if captcha exists and is unexpired.

Orphan instances

FromRow ThreadInfo

Instance details

Methods

fromRow :: RowParser ThreadInfo

FromRow Post

Instance details

Methods

fromRow :: RowParser Post

FromRow BoardConstraints

Instance details

Methods

fromRow :: RowParser BoardConstraints

FromRow BoardInfo

Instance details

Methods

fromRow :: RowParser BoardInfo

FromRow AccountInfo

Instance details

Methods

fromRow :: RowParser AccountInfo