Imageboard.FileUpload

Documentation

saveFile :: MonadIO m => File -> FileData -> FilePath -> ExceptT Text m File

Try to save file, recognize its size and dimensions and generate thumbnail for it.

tryMkFile :: FileData -> Either Text (File, FilePath)

Validate file send by the user. This function recognizes format of sent file and its size and fails if they do not match predefined requirements. Otherwise, it returns new file and filename.

type FileData = FileInfo ByteString

Type representing file as received by HTTP server.