Imageboard.Pages.Common

Documentation

renderFlags :: ThreadFlags -> Html

Render flags that apply to a thread.

space :: Html

A space character.

commonHtml

Arguments

:: Text
Title

-> [Board]
List of boards

-> Html
Inner HTML

-> Html

Render HTML with title and navbar that includes list of boards.

addTopBottom :: Html -> Html

Add navigation links that bring to top and bottom of the page.

threadForm :: Board -> Html

A form for creating threads on given board.

replyForm :: Board -> Int -> Html

A form for creating replies to thread specified by board and number.

fileThumbLink :: File -> Text

formatDate :: UTCTime -> String

Format date to YYYY-MM-DD HH:MM:SS.