Imageboard.Types

Documentation

data Post

Constructors

Post

Fields

loc :: PostLocation

date :: UTCTime

UTC date of post creation.

content :: PostStub

file :: Maybe File

File if attached.

data PostStub

Constructors

Stub

Fields

author :: Text

email :: Text

subject :: Text

text :: Text

Post body.

data File

Constructors

File

Fields

filename :: Text

Exact file name, including extension.

ext :: FileType

File format.

size :: Int

Size in bytes.

dim :: Maybe Dimensions

Dimensions if they apply.

data FileType

Data type representing supported file formats.

Constructors

JPG

PNG

GIF

WEBM

MP4

MP3

OGG

PDF

EPUB

SWF

isImage :: FileType -> Bool

Whether given file format may store an image.

isAudio :: FileType -> Bool

Whether given file format may store an audio.

data Dimensions

Dimensions of raster image or video.

Constructors

Dim

Fields

width :: Int

height :: Int

data ThreadFlags

Constructors

Flags

Fields

sticky :: Bool

Whether thread is sticked to the top of the board.

lock :: Bool

Whether thread is closed for replies.

autosage :: Bool

Whether thread can be bumped to the top of the board.

cycle_ :: Bool

Whether old replies will be deleted will after limit of max replies is reached.

data ThreadInfo

Constructors

ThreadInfo

Fields

lastBump :: UTCTime

flags :: ThreadFlags

replyCount :: Int

Number of replies in thread.

data ThreadHead

Constructors

ThreadHead

Fields

opPost :: Post

Opening post of a thread.

opInfo :: ThreadInfo

data Thread

Thread as a whole.

Constructors

Thread

Fields

op :: ThreadHead

replies :: [Post]

data PostLocation

Constructors

PostLocation

Fields

board :: Board

Name of board post was posted on.

number :: Int

Post number, unique on board.

parent :: Maybe Int

Number of parent post.

type Board = Text

Unique name of a message board.

data BoardInfo

Information specific to board.

Constructors

BoardInfo

Fields

name :: Board

title :: Text

subtitle :: Text

data BoardConstraints

Constructors

Constraints

Fields

isLocked :: Bool

Whether board is closed for new threads.

minLen :: Int

Minimum length for a post.

maxLen :: Int

Maximum length for a post.

maxNewLines :: Int

Maximum number of new lines in a post.

maxReplies :: Int

Maximum number of replies in a thread.

maxThreads :: Int

Maximum number of threads on the board.

type Username = Text

data AccountInfo

Constructors

AccountInfo

Fields

user :: Username

accountCreated :: UTCTime

role :: Role

data Role

Constructors

Admin

Moderator

type SessionKey = Text

Random token used for authentication.