Regex.PCRE2

Description

PCRE2 bindings for Data.Text. |

Documentation

data RegexReplace

Constructors

REReplace

Fields

pattern, replacement :: Text

gsub :: [RegexReplace] -> Text -> Text

Match patterns and do substitutions. This function will try to compile given patterns and apply given substitutions in order they appear in input list. It may throw runtime error if submitted match pattern is incorrect.